analytics_service/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiosqlite==0.19.0
alembic==1.12.1

# Cache and messaging
redis==5.0.1
aioredis==2.0.1

# Elasticsearch
elasticsearch==8.11.1

# HTTP client
httpx==0.25.2
requests==2.31.0

# Analytics and ML
pandas==2.1.3
numpy==1.25.2
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2

# Authentication and security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Templates
Jinja2==3.1.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
ruff==0.1.6
mypy==1.7.1

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0
